requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
pytest==7.1.2
pyzbar==0.1.9
six==1.16.0
tomli==2.0.1
validators==0.20.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
